requirements.txt
google.generativeai
flask
torch
gunicorn
